Add Energy, Momentum, and Mass macros
Also add the `xspace` package and `amsmath`. `xspace` adds a macro,
`\xspace` that tries to detect if that macro should include a space or
not (for example, if it is in the middle of a sentence or at the end).
`amsmath` includes many useful math formatting options, including
`\text` for embedding text in math.